Rename & refactor MtAmazonDynamoDbContextProviderImpl
- Rename the impl class not to be confusing
- Refactor the internal ThreadLocal value
